requirements.txt
aiohttp
aiohttp-cors
numpy
python-socketio==5.11.3
